fix some struct code
