+++ components/languages.js
@@ -1,22 +1,14 @@
 import { countries_data } from "../data/countries_data.js";
+import { getLangagesCount, getTenMostSpokenLanguages } from "../utils/utils.js";
 
 export function tenMostSpokenLanguages() {
   const container = document.createElement("div");
   container.classList.add("box");
-  const allLanguages = countries_data
-    .map((country) => country.languages)
-    .flat();
-
-  const languageCounts = allLanguages.reduce((acc, language) => {
-    acc[language] = (acc[language] || 0) + 1;
-    return acc;
-  }, {});
+  const languageCounts = getLangagesCount(countries_data);
 
   const topTenLanguages = document.createElement("div");
   topTenLanguages.classList.add("items-container");
-  const languageArray = Object.entries(languageCounts);
-  const sortedLanguages = languageArray.sort((a, b) => b[1] - a[1]);
-  const sortedTenLanguages = sortedLanguages.slice(0, 10);
+  const sortedTenLanguages = getTenMostSpokenLanguages(languageCounts);
 
   for (const element of sortedTenLanguages) {
     const languageElement = document.createElement("div");
