+++ components/subtitle.js
@@ -4,5 +4,4 @@ export function subtitle() {
   const numberOfCountries = countries_data.length;
   const subtitle = document.querySelector(".subtitle");
   subtitle.textContent = `Currently, we have ${numberOfCountries} countries`;
-  return subtitle;
 }
