+++ index.html
@@ -23,6 +23,7 @@
         <div class="graph-buttons">
           <button class="population">Population</button>
           <button class="languages">Languages</button>
+          <button class="test">test</button>
         </div>
         <h4 class="graph-title"></h4>
         <div class="graphs">
